components/loan-management/scripts/data_migrate/requirements.txt
requests
pyyaml
python-dotenv
dacite
urllib3
pandas
numpy
